+++ resources/snippet.php
@@ -11,7 +11,7 @@ $mail_body = '<html>
 <p>Dear '.$username.'<br><br>Thank you for registering, please click on the link below to
 	confirm your email address</p>
 <p><a href="http://auth.dev/activate.php?id='.$encode_id.'"> Confirm Email</a></p>
-<p><strong>&copy;2016 ICT DesighHUB</strong></p>
+<p><b>&copy;2016 ICT DesighHUB</b></p>
 </body>
 </html>';